+++ Dracula/Cards/Secrets/PlaceholderSecretCard.cs
@@ -13,7 +13,7 @@ internal sealed class PlaceholderSecretCard : SecretCard, IDraculaCard
 			CardType = MethodBase.GetCurrentMethod()!.DeclaringType!,
 			Meta = new()
 			{
-				deck = ModEntry.Instance.DraculaDeck.Deck,
+				deck = ModEntry.Instance.SpellDeck.Deck,
 				rarity = Rarity.common,
 				dontOffer = true,
 				unreleased = true